Crate ez_jsonrpc_types

A transcription of types from the JSON-RPC 2.0 Specification.

When quoted, the specification will appear as blockquoted text, like so.

Design

All structs are owned (i.e, there is no borrowing of data from the Deserializer), to facilitate ergonomics.

Appearances of dynamic JSON Values are parameterised out, to allow deferred serialization using, i.e RawValue.

Re-exports

pub use map::Map;

Modules

map

A map of Strings to values (typically Value).

params

Extra support for parameter (de)serialization

Structs

Error

A JSON-RPC 2.0 error object.

Request

A JSON-RPC 2.0 request object.

Response

A JSON-RPC 2.0 response object.

Enums

Id

See Request::id.

MaybeBatchedRequest

To send several Request objects at the same time, the Client MAY send an Array filled with Request objects.

MaybeBatchedResponse

A response to a MaybeBatchedRequest.

RequestParameters

If present, parameters for the rpc call MUST be provided as a Structured value. Either by-position through an Array or by-name through an Object.